Inform user when a new directory has been made.
I have seen that a number of users have been confused by a featured introduced in an earlier version.  If 'yo famous' is run in a directory that already has files, a new directory is made and everything is put in there.  While this should be expected behavior, some users were confused to find nothing worked when running grunt.  Users will now be prompted to also cd into the newly created directory.